[Kernel][FS] Add open callback for filesystems
