+++ Kernel/Drivers/Base/CharacterDevice.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <Filesystem/Base/FileDescriptor.hpp>
+
 #include <Macaronlib/Common.hpp>
 #include <Macaronlib/String.hpp>
 
@@ -7,6 +9,7 @@ namespace Kernel::Drivers {
 
 class CharacterDevice {
 public:
+    CharacterDevice() = default;
     CharacterDevice(const String& name)
         : m_name(name)
     {
@@ -14,7 +17,7 @@ public:
 
     const String& name() const { return m_name; }
 
-    virtual void open() { }
+    virtual void open(FS::FileDescriptor& fd) { }
     virtual uint32_t write(uint32_t offset, uint32_t size, void* buffer) { return 0; }
     virtual uint32_t read(uint32_t offset, uint32_t size, void* buffer) { return 0; }
     virtual bool can_read(uint32_t offset) { return false; }
@@ -23,7 +26,7 @@ public:
     virtual bool ioctl(uint32_t request) { return false; };
 
 protected:
-    String m_name;
+    String m_name {};
 };
 
 }
